requirements.txt
wmi
pystray
Pillow
pywin32